Various changes for general cleanup and to support the upcoming svnci.el.
Thorough font-locking regexp's based on the specifications of the
status command.  They are also now flexible enough to handle both
commit and status output, which differ slightly.

Help text is now generated based on the actual keymaps on the fly

Modularized the code for editing buffers without marking them modified
and the code for making regions of the buffer immutable.


git-svn-id: http://awakening.csail.mit.edu/svn/amthrax/sys/elisp@1689 ee464a63-62dc-0310-8e92-97079e79a51f