Made tooltips computed on the fly
Fixed up the regexp's a bit more.  Ignored files are now understood
and external files are shown in the default face instead of the
conflict face.

Tooltip and keymap properties are now done with an overlay instead of
text properties so that they don't do strange things when the text
gets moved around.

Visit added/committed/etc files instead of diff'ing them

Inhibit flyspell on the info block


git-svn-id: http://awakening.csail.mit.edu/svn/amthrax/sys/elisp@1691 ee464a63-62dc-0310-8e92-97079e79a51f